+++ api/v1/linstorsatelliteconfiguration_types.go
@@ -74,6 +74,10 @@ type LinstorSatelliteConfigurationSpec struct {
 	// +kubebuilder:validation:Optional
 	DeletionPolicy DeletionPolicy `json:"deletionPolicy,omitempty"`
 
+	// +kubebuilder:validation:Optional
+	// +nullable
+	EvacuationStrategy *EvacuationStrategy `json:"evacuationStrategy,omitempty"`
+
 	// Template to apply to Satellite Pods.
 	//
 	// The template is applied as a patch to the default resource, so it can be "sparse", not listing any